CONTRIBUTING: Add legal section and explain DCO signoff
[1] currently covers the DCO and Signed-off-by in the community
repository, and this taxonomy repository links that file with:

  Please read the [community contribution guide](https://github.com/instruct-lab/community/blob/main/CONTRIBUTING.md) first for general practices for the InstructLab community.

But a link floating with the 'main' branch like that is not very
reliable, and in a future world where the organization moves to a DCO
1.2 or similar, it would be hard to determine which version a
contributor was signing in their commit.  By including the name and
version of the DCO locally in this repository, that ambiguity is
removed; the contributor was signing the version mentioned in this
repository when they made their commit.

I'd prefer to have inlined the complete DCO text in this repository,
instead of linking out to https://developercertificate.org/ .  But the
licensing of the DCO itself is more restrictive than Apache-2:

  Copyright (C) 2004, 2006 The Linux Foundation and its contributors.

  Everyone is permitted to copy and distribute verbatim copies of this
  license document, but changing it is not allowed.

so I'm following the example set by [1] and linking out to external text.

[1] also chooses some external links I don't understand for both the
DCO text and the fact that the Linux community uses it, and I'm
replacing those with what I consider more canonical links.

I'm also linking LICENSE and mentioning Apache-2 to make the meaning
of DCO (b)'s "appropriate open source license" more explicit than just
"we happened to have the Apache-2 text in LICENSE".

[1]: https://github.com/instruct-lab/community/blob/b64fc4052c0eee427efc051bd73f06e85b1b319b/CONTRIBUTING.md#legal

Signed-off-by: W. Trevor King <wking@tremily.us>