Add a skill to generate allow rule from SELinux denial
Without this patch the following was observed:
>>> Generate allow rule from the following SELinux denial: "type=AVC msg=audit(1226882736.442:86): avc:  denied
{ getattr } for  pid=2427 comm="httpd" path="/var/www/html/file1" dev=dm-0 ino=284133 scontext=unconfined_u:system_r:httpd_t:s0
tcontext=unconfined_u:object_r:samba_share_t:s0 tclass=file"

Output:
To generate an allow rule for the given SELinux denial, we need to specify the SELinux context for the file,
which is not provided in the original message. However, I can provide you with a general template for creating an
allow rule.

To create an allow rule, you can use the command `semanage`, which is a command-line tool for managing SELinux
policies. Here's a template for creating an allow rule for the given denial:

```bash
semanage fcontext -a -t samba_share_t "/var/www/html/file1(/.*)?"
```
This command will add a new context to the SELinux policy, allowing the specified file to be accessed by the
httpd process. However, please note that this command assumes that the file is located in the `/var/www/html/`
directory and has the name `file1`. If the file is located in a different directory or has a different name, you
will need to adjust the path in the `semanage fcontext` command accordingly.

After applying this patch the expected output should looks like:

>>> Generate allow rule from the following SELinux denial: "type=AVC msg=audit(1226882736.442:86): avc:  denied
{ getattr } for  pid=2427 comm="httpd" path="/var/www/html/file1" dev=dm-0 ino=284133 scontext=unconfined_u:system_r:httpd_t:s0
tcontext=unconfined_u:object_r:samba_share_t:s0 tclass=file"

Output:
To generate an SELinux allow rule for the given denial, you can manually construct the rule based on the denial
information: allow httpd_t samba_share_t:file getattr;

Signed-off-by: Lukas Vrabec <lvrabec@redhat.com>