quantitive: recognize function solutions given in context
An interesting exchange happens with MT-Bench math#120 and several models:

> User: Given that f(x) = 4x^3 - 9x - 14, find the value of f(2).
>
> Assistant:
> Okay, let’s solve this step-by-step:
>
> Given: f(x) = 4x^3 - 9x - 14
>
> We want to find f(2), which means we substitute 2 for x: f(2) = 4(2)^3 - 9(2) - 14
>
> Simplify: f(2) = 4(8) - 18 - 14 = 32 - 18 - 14 = 32 - 32 = 0
>
> Therefore, the value of f(2) is 0.
>
> User follow-up: Find x such that f(x) = 0.

Of course, at this point, a human would notice that the assistant should have
recognized from its first answer that x = 2 is one solution (in fact it is the
only real solution).

So this is an example where perhaps a model could perform better on the MT-Bench
if it could be thought to recognize when a solution for a function has been given
rather than attempting to "do the math" to work out the solution itself.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>