skill: Add the ability to generate tldr-inspired usage advice
Previously, the llm was inclined to give verbose answers about how to
use a cli with few actual cli examples. While this is good when you're
trying to learn a lot about a tool, it isn't helpful when you just need
a "quickstart" version of tool usage.

This change teaches the llm to respond in a manner closer to what we'd
expect from a tool like tldr, though it is still a bit more verbose.

Signed-off-by: Jacob Callahan <jacob.callahan05@gmail.com>