RFC: WIP: add dynamic cyrillic translit linguistic skill
This is an attempt to add support for a Cyrillic-to-Latin
transliteration skill that would also be language aware.

This is not validated and not working well with the default prompt
defined in cli repo. (The prompt produces new instructions that are not
retaining the transliteration scheme.)

Whether we need to use a model to generate variations of the generated
instructions is itself a topic for exploration. (Perhaps we could live
with feeding the seed samples generated by the skill directly into
fine tuning.) But if so, additional changes may be required on `cli`
side to e.g. allow to use modified (or completely new?) prompts if
needed for a particular skill.

This is work-in-progress and is posted as a discussion starter.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>