Add a skill for ROT13
This adds a grounded compositional editing skill to teach ROT13 to the model.
It is grounded because instructions/context about what ROT13 is and how to apply
it are provided as `context` in each example.

However, most current models are not capable of actually understanding and
applying these instructions. Therefore, synthetic training sets generated for
this skill are likely to be inaccurate.

At this time, this skill can be more useful as a benchmark / test for the
process itself than as an actual skill.

Signed-off-by: Pep Turró Mauri <pep@redhat.com>