Freeform Composition Skill: more limerick examples
Add limericks and non-limericks on some random topics, with the
idea being that the model will have data to distinguish the difference
when asked to generate limericks, because the initial response based
on the current 1 sample limerick is very disappointing.

All limericks in this commit are written by me.
All non-limericks are "limerick" responses from the model
merlinite-7b-Q4_K_M.gguf

Signed-off-by: Brian Rosmaita <rosmaita.fossdev@gmail.com>