Skill to convert text to Bionic Reading
Bionic Reading it's a modification of a text to increase reading speed.
It is basically formatting with bold the initial letters of each words
to stand out.

This skill includes 20 examples of this conversion using markdown to
format the text so it can do bold.

Bionic Reading page: https://bionic-reading.com/

Signed-off-by: Gorka Eguileor <geguileo@redhat.com>