README: Mention compute infrastructure as a fork cost
This section already started out with:

  The ability to contribute to a large language model (LLM) has been
  difficult in no small part because it is difficult to get access to
  the necessary compute infrastructure.

My new addition points out that that cost also applies to forks, to
underline the value of things like centralized frequent model
retraining.

And forking can dilute/fracture the community, so I'm also callling
out the value of that community in sharing the
maintenance/contribution load.

But sometimes folks have work that they cannot upstream, either
because it does not align with upstream-maintainer goals, or because
they need to keep it internal.  My new paragraph points out that this
is possible, but surrounds that possibility with reminders of the
costs, so folks weighing those trade offs can make a more informed
decision.

Signed-off-by: W. Trevor King <wking@tremily.us>