Add skill to recommed good Ansible tasks
This skill provides examples of good Ansible tasks aligned to Ansible
best practices but also considering Playbooks performance and
readability. This skill needs more knowledge and will likely be
incremented in the future but for now it was already noticed some
improvements in the answers after the inclusion of this Ansible skill.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>