schema: We add attribution to the taxonomy schema
All examples must contain an attribution key. The lint check is updated
to test for the presence of the attribution key and its subkeys.

attribution is an array of mappings where each mapping has a source and
license key. source is a URL to 3rd party information or `self-authored`
if the contributor authored the information. license is the SPDX
license identifier of the 3rd party information or `Apache-2.0` (repo
license) if the contributor authored the information.

Fixes https://github.com/instruct-lab/taxonomy/issues/182

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>