docs: correct merge conflict
Recent churn in README.md caused a merge conflict between #507 (moved
the '## YAML Format' section earlier in the document) and #514
(altered content in the '## YAML Format' section in its older
location), which resulted in the README.md having two inconsistent
YAML Format sections.  Recombine the two '## YAML Format' sections
into one, with the intended text changes from the latter in the
correct location of the former.

Similarly, the original proposal for #514 was to intentionally remove
a fourth pun from the example, leaving only 3; but the addition of
'attribution:' keys in an earlier commit meant that the example was
changed in the meantime to list 5 puns to match the underlying file;
the merge conflict was incorrectly resolved by leaving 5 puns intact.
The correct merge resolution is to get back to an intentional 3 puns
in the README.

Fixes: #507
Fixes: #514
Signed-off-by: Eric Blake <eblake@redhat.com>