Recognize when statements are referring to the same place.
I.e., if m different statements are made about a given space having n things,
then the total number of things in the space is n and not m*n.

For example, if I ask the question "Brooke lives with three cats. Alice lives
with Brooke. How many cats does Alice live with?"  The model recognizes that
they live together, and responds that Alice lives with 3 cats as expected.

However, if I ask the similar question "Brooke lives with three cats. Alice
lives with three cats. Alice and Brooke live together in the same house. How
many cats are in the house?" The model does not infer that we are talking about
the same three cats and responds with 3+3=6 cats.

I have asked several similar questions (see `qna.yaml` file for a selection) and
in all cases the response is similar to the unexpected response above.

I initially trained the model with the `qna.yaml` from pr #365 to see if it
would help, but it didn't. I then tried retraining with both the file from pr
number of variables, I trained the model with only my file, but that didn't help
either. I also tried a couple different ways to word my answers but those also
didn't help.

Unfortunately, my attempts to train the model were not successful. However, I
am submitting this as a draft in case it provides some useful information or if
anyone has suggestions for how to train the model better.

I was curious about whether these questions were reasonable, so I asked them to
ChatGPT. In all cases, ChatGPT responded with the answers that I was expecting
with a good explanation of the rationale.

Two examples are given below.

  Q1:
  Andrew and Bob live on the same street. There are two stop signs on Bob's
  street. There are two stop signs on Andrew's street. How many stop signs are
  there?

  Q2:
  Alice and Bob live together. Alice lives with 2 dogs. Bob lives with two dogs.
  How many dogs are in the house?

Response from the original model

  A1:
  There are two stop signs for each of the two people, so there are 2 x 2 = 4
  stop signs in total. The answer is 4

  A2:
  Alice and Bob each have 2 dogs, so between them, they have 2 dogs + 2 dogs = 4
  dogs. Since they live together, there are 4 dogs in the house.

Response from the fine-tuned model

  A1':
  There are 4 stop signs in total. Since Bob and Andrew both live on the same
  street, and there are two stop signs on each of their streets, there must be
  2 x 2 = 4 stop signs in total.

  A2':
  There are four dogs in the house because both Alice and Bob have two dogs
  living with them in the same house.

Signed-off-by: Andre Fredette <afredette@redhat.com>