schema: Define JSON schema for taxonomy yaml
We use JSON Schema which has broad support including in vscode. This PR
includes the schema for compositional skills (without attribution) as
well as prototype schema for knowledge and metadata. The check yaml
script is changed to use the check-jsonschema command to validate
the compositional skill qna.yaml files against the schema.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>