schema: Move schema files into a versioned folder
To prepare for future updates to the schema, we use a versioned folder.
Then downstream users, such as cli, can use the schema file from the
versioned folder(s) which the user understands. This will also allow
users to support and validate documents from older schema versions.

We also add the version property to the schema. This is supported by
cli from https://github.com/instruct-lab/cli/pull/417. The version
property is not required for version 1.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>