Add context for non-limericks.
Instead of giving seed questions for not-a-limerick, provide
seed questions to explain why a poem given in context is not
a limerick.

The four-line poems given in context are "limerick" responses from
the model merlinite-7b-Q4_K_M.gguf.  The other non-limericks given
in context were adapted by me from "limerick" responses from that
same model.

Signed-off-by: Brian Rosmaita <rosmaita.fossdev@gmail.com>