knowledge: Pivot to Git submodules
As documented in [1,2].  This way we can use that off-the-shelf
tooling for linking dependent Git repositories, instead of building
something similar on our own.  I'm dropping 'repo' and 'commit' now
that that portion has been offloaded to submodules.

There's no upstream-Git analog to pattern, although recent versions do
have an experimental sparse-checkout [3].  But the files will be
pulled down by the clone either way, so it doesn't seem like a big
deal to land them all on the filesystem, and then use the regular
expressions to iterate over the files you actually want to feed the
model while training.

document:
  pattern:
  - regular expression 1
  - regular expression 2
  - ...

seemed a bit chatty, now that we're only holding an array of regular
expressions, so I'm flattening to a single:

documents:
- regular expression 1
- regular expression 2
- ...

[1]: https://git-scm.com/docs/gitsubmodules
[2]: https://git-scm.com/book/en/v2/Git-Tools-Submodules
[3]: https://git-scm.com/docs/git-sparse-checkout

Signed-off-by: W. Trevor King <wking@tremily.us>