check_yaml: Update for better command line use including yamllint
The script now calls yamllint.

It is also updated to work better on
the command line where the user may pass paths not relative to a
taxonomy folder. We compute the schema name from the first path
element or from the presence/absence of the document element.

We remove the need for the matcher by emitting github workflow messages.
Since the script is in control of the messages emitted, we can emit
them in the necessary format rather than having to use a matcher to get
the workflow to match.

Fixes https://github.com/instruct-lab/taxonomy/issues/678

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>