Rewrite check_yaml script in python
This allows us more control over the handling of validation messages.
In particular, we improve the minItems=5 message for seed_examples to
be shorter.

The script can now also be configured by some env vars.

Unfortunately, there is no decent yq python package which supports the
needed expression. So we must use the yq command.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>