stale: Update configuration for issue staleness
Since the stale bot is configured, we also need to configure for issues.
When an issue is marked stale, we want a comment added to alert those
following the issue. We also want to configure the days rather than use
the default.

Finally, we exempt PRs labeled triage-approved from being marked stale
to allow the necessary time for such PRs to go through the backend
process.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>