Even after yamllint runs we have more work
It seems that we have more things that need to be checked apart
from just yamllinting. This is a starter script to validate
after it's good yaml that the expected things are in the actual
yaml file.
It just exits(1) when something is missing or None, which
`created_by:` seems to be missed a lot.

Signed-off-by: JJ Asghar <awesome@ibm.com>