workflow: Handle when a pull request does not contain qna.yaml (#1112)
Some pull requests are not about skill or knowledge contributions. We
change to always upload an artifact so the insight workflow wont fail
because it cannot download the artifact. In the case were the PR does
not contain qna.yaml, we pass the pull request number 0 so the insight
workflow can avoid invoking the insight analysis.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>