Update and rename qna.txt to qna.yaml
I *hopefully* fixed the yaml correctly. I looked at the taxonomy example and realized I need to indent the answer: section.  I did all of this at the Red Hat Summit in a lab on the show floor on an unfamiliar laptop in a hurry.  I am wearing my instruct lab hoodie right now as I try to fix this. 

Signed-off-by: Eric Adams <eric.adams@intel.com>