schema: Define a relaxed schema as v1 and move current schema to v2 (#769)
We introduce a relaxed schema as v1 which will accept all existing
taxonomy files in the main branch of the taxonomy repo. The current v1
schema is renamed to v2 and will be used for future contributions to the
taxonomy repo.

The check_yaml script is updated to allow a schema version to be
specified. When the schema-version option is not specified, the script
will use the latest schema version from the schema base. If the
schema-version is specified as a non-positive integer, then the script
will use the version key from the YAML document for the schema version
to use.

The script is also changed to only check for attribution.txt when the
schema version is higher than 1.

See https://github.com/instructlab/instructlab/issues/989

Depends on https://github.com/instructlab/schema/pull/11

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>