Removed ibm_redbooks
Removing the ibm_redbooks until the documentation can be linked
We do not know where these questions come from, and if we can use
the ibm redbooks due to the lack of the `document:` portion being
not set up correctly.

We will remove this qna.yaml for the time being, but keep the directory
if someone can get it submitted correctly for the future.

/cc @russellb @luke-inglis

Signed-off-by: JJ Asghar <awesome@ibm.com>