Update KNOWLEDGE_GUIDE.md
Reorder bulleted list in topics to avoid, clarify jokes & poems language.

I am bypassing regular review process for these minor changes because I am on deadline to get an article on InstructLab to a member of the press and need to link to this documentation. 

I acknowledge this is not best practice and do not intend to make this a habit.

Signed-off-by: Leslie Hawthorn <lhawthor@redhat.com>